improve generator
