+++ .gitignore
@@ -462,3 +462,4 @@ blockfrost-client/
 /samples/client/src/Blockfrost.Api/Blockfrost.Api.csproj
 appsettings.test.json
 /src/Blockfrost.Client
+/samples/unity/BlockfrostJump
