+++ samples/blockfrost-dotnet-samples.sln
@@ -26,6 +26,12 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		switcher.json = switcher.json
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlockfrostWithAzure", "net5\BlockfrostWithAzure\BlockfrostWithAzure.csproj", "{ED16C384-A0CE-46FB-A487-6B27E6AC7B38}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Blockfrost.Api", "..\src\Blockfrost.Api\Blockfrost.Api.csproj", "{AAF52E1C-37F5-46FA-B43A-768DB07175F1}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Blockfrost.Extensions", "..\src\Blockfrost.Extensions\Blockfrost.Extensions.csproj", "{D87E0100-FD22-4FDE-8E81-2087E7D37F05}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -96,6 +102,42 @@ Global
 		{0E102E56-2529-41D4-8327-2CAC963367E7}.Release|x64.Build.0 = Release|Any CPU
 		{0E102E56-2529-41D4-8327-2CAC963367E7}.Release|x86.ActiveCfg = Release|Any CPU
 		{0E102E56-2529-41D4-8327-2CAC963367E7}.Release|x86.Build.0 = Release|Any CPU
+		{ED16C384-A0CE-46FB-A487-6B27E6AC7B38}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{ED16C384-A0CE-46FB-A487-6B27E6AC7B38}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{ED16C384-A0CE-46FB-A487-6B27E6AC7B38}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{ED16C384-A0CE-46FB-A487-6B27E6AC7B38}.Debug|x64.Build.0 = Debug|Any CPU
+		{ED16C384-A0CE-46FB-A487-6B27E6AC7B38}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{ED16C384-A0CE-46FB-A487-6B27E6AC7B38}.Debug|x86.Build.0 = Debug|Any CPU
+		{ED16C384-A0CE-46FB-A487-6B27E6AC7B38}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{ED16C384-A0CE-46FB-A487-6B27E6AC7B38}.Release|Any CPU.Build.0 = Release|Any CPU
+		{ED16C384-A0CE-46FB-A487-6B27E6AC7B38}.Release|x64.ActiveCfg = Release|Any CPU
+		{ED16C384-A0CE-46FB-A487-6B27E6AC7B38}.Release|x64.Build.0 = Release|Any CPU
+		{ED16C384-A0CE-46FB-A487-6B27E6AC7B38}.Release|x86.ActiveCfg = Release|Any CPU
+		{ED16C384-A0CE-46FB-A487-6B27E6AC7B38}.Release|x86.Build.0 = Release|Any CPU
+		{AAF52E1C-37F5-46FA-B43A-768DB07175F1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AAF52E1C-37F5-46FA-B43A-768DB07175F1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AAF52E1C-37F5-46FA-B43A-768DB07175F1}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{AAF52E1C-37F5-46FA-B43A-768DB07175F1}.Debug|x64.Build.0 = Debug|Any CPU
+		{AAF52E1C-37F5-46FA-B43A-768DB07175F1}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{AAF52E1C-37F5-46FA-B43A-768DB07175F1}.Debug|x86.Build.0 = Debug|Any CPU
+		{AAF52E1C-37F5-46FA-B43A-768DB07175F1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AAF52E1C-37F5-46FA-B43A-768DB07175F1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AAF52E1C-37F5-46FA-B43A-768DB07175F1}.Release|x64.ActiveCfg = Release|Any CPU
+		{AAF52E1C-37F5-46FA-B43A-768DB07175F1}.Release|x64.Build.0 = Release|Any CPU
+		{AAF52E1C-37F5-46FA-B43A-768DB07175F1}.Release|x86.ActiveCfg = Release|Any CPU
+		{AAF52E1C-37F5-46FA-B43A-768DB07175F1}.Release|x86.Build.0 = Release|Any CPU
+		{D87E0100-FD22-4FDE-8E81-2087E7D37F05}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D87E0100-FD22-4FDE-8E81-2087E7D37F05}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D87E0100-FD22-4FDE-8E81-2087E7D37F05}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{D87E0100-FD22-4FDE-8E81-2087E7D37F05}.Debug|x64.Build.0 = Debug|Any CPU
+		{D87E0100-FD22-4FDE-8E81-2087E7D37F05}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{D87E0100-FD22-4FDE-8E81-2087E7D37F05}.Debug|x86.Build.0 = Debug|Any CPU
+		{D87E0100-FD22-4FDE-8E81-2087E7D37F05}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D87E0100-FD22-4FDE-8E81-2087E7D37F05}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D87E0100-FD22-4FDE-8E81-2087E7D37F05}.Release|x64.ActiveCfg = Release|Any CPU
+		{D87E0100-FD22-4FDE-8E81-2087E7D37F05}.Release|x64.Build.0 = Release|Any CPU
+		{D87E0100-FD22-4FDE-8E81-2087E7D37F05}.Release|x86.ActiveCfg = Release|Any CPU
+		{D87E0100-FD22-4FDE-8E81-2087E7D37F05}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -106,6 +148,7 @@ Global
 		{063DA81E-4C0D-4DF4-9E68-D10758F3CCB6} = {3DDB4AE1-0D3A-4883-AAB2-FC9DDD60FAD1}
 		{043493AB-CB6B-4A2E-97FA-245B57512068} = {932CEC78-4C1C-4C60-A513-0FAFFBE736ED}
 		{0E102E56-2529-41D4-8327-2CAC963367E7} = {932CEC78-4C1C-4C60-A513-0FAFFBE736ED}
+		{ED16C384-A0CE-46FB-A487-6B27E6AC7B38} = {932CEC78-4C1C-4C60-A513-0FAFFBE736ED}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {1712CD06-8E84-4B5F-BC51-F0DCAD33A545}
