+++ samples/net5/BlockfrostWithAzure/BlockfrostWithAzure.csproj
@@ -2,20 +2,19 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>net5.0</TargetFramework>
-    <UserSecretsId>e4b04c6d-2256-4eec-a45e-bfe2bb724686</UserSecretsId>
+    <UserSecretsId>8d368a40-e50f-478b-bb18-34bd40573fc7</UserSecretsId>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="CardanoSharp.CBOR" Version="0.1.0" />
-    <PackageReference Include="CardanoSharp.Wallet" Version="0.2.6" />
     <PackageReference Include="Microsoft.Extensions.Configuration.UserSecrets" Version="5.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="5.0.0" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="5.0.2" />
   </ItemGroup>
   <ItemGroup>
     <None Update="appsettings.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\..\src\Blockfrost.Api\Blockfrost.Api.csproj" />
+    <ProjectReference Include="..\..\..\src\Blockfrost.Extensions\Blockfrost.Extensions.csproj" />
   </ItemGroup>
 </Project>
