+++ samples/net5/BlockfrostWithAzure/appsettings.json
@@ -1,0 +1,12 @@
+﻿{
+  //"Blockfrost": {
+  //  "Blockfrost.Net.Sdk": {
+  //    "Network": "testnet"
+  //  }
+  //}
+  "Blockfrost": {
+    "Name": "Blockfrost.Net.Sdk",
+    "ApiKey": "kL2vAF27FpfuzrnhSofc1JawdlL0BNkh",
+    "Network": "testnet"
+  }
+}
