+++ samples/netcore31/console/ConsoleApp1/ConsoleApp1.csproj
@@ -4,7 +4,9 @@
     <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Blockfrost.Api" Version="0.1.1" />
     <PackageReference Include="System.Text.Json" Version="5.0.2" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\..\src\Blockfrost.Api\Blockfrost.Api.csproj" />
   </ItemGroup>
 </Project>
