+++ samples/netstandard21/console/ConsoleApp1/ConsoleApp1.csproj
@@ -4,8 +4,10 @@
     <TargetFramework>netstandard2.1</TargetFramework>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Blockfrost.Api" Version="0.1.1" />
     <PackageReference Include="System.ComponentModel.Annotations" Version="5.0.0" />
     <PackageReference Include="System.Text.Json" Version="5.0.2" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\..\src\Blockfrost.Api\Blockfrost.Api.csproj" />
   </ItemGroup>
 </Project>
