+++ src/Blockfrost.Api/Options/IBlockfrostProject.cs
@@ -1,0 +1,10 @@
+﻿namespace Blockfrost.Api.Options
+{
+    public interface IBlockfrostProject
+    {
+        string ApiKey { get; set; }
+        string Name { get; set; }
+        string Network { get; set; }
+        string ConnectionLimit { get; set; }
+    }
+}
