+++ src/Blockfrost.Extensions/AzureKeyVault/ServiceCollectionExtension.cs
@@ -1,0 +1,12 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+
+namespace Blockfrost.Extensions.AzureKeyVault
+{
+    public static class ServiceCollectionExtension
+    {
+        public static IServiceCollection AddBlockfrost(this IServiceCollection services, string azureSecretName)
+        {
+            return services;
+        }
+    }
+}
