+++ test/Blockfrost.Api.Tests/Extensions/AppSettingsConfigurationTests.cs
@@ -11,11 +11,25 @@ namespace Blockfrost.Api.Tests.Extensions
     [TestCategory(nameof(Extensions))]
     public class AppSettingsConfigurationTests : AServiceTestBase
     {
-
         [ClassInitialize]
         public static void Setup(TestContext context)
         {
             ConfigureEnvironment(Constants.PROJECT_NAME_TESTNET, context);
+        }
+
+        [TestMethod]
+        public void BlockfrostConfigSection_WithoutDictionary()
+        {
+            var config = new ConfigurationBuilder()
+                            .AddJsonFile($"appsettings.single.json", optional: false, reloadOnChange: true)
+                            .AddJsonFile("appsettings.json", optional: true, reloadOnChange: true)
+                            .Build();
+
+            IServiceCollection services = new ServiceCollection();
+            var provider = services.AddBlockfrost(config).BuildServiceProvider();
+            var ledgerService = provider.GetRequiredService<ILedgerService>();
+            Assert.IsNotNull(ledgerService);
+            Assert.AreEqual("testnet", ledgerService.Network);
         }
 
         [TestMethod]
