+++ test/Blockfrost.Api.Tests/appsettings.single.json
@@ -1,0 +1,18 @@
+﻿{
+  // uncommented to use this instead of the system's environment variable
+  "NETCORE_ENVIRONMENT": "staging",
+
+  // Simple configuration
+  "Name": "Blockfrost.Net.Sdk-cli",
+  "ConnectionLimit": 10,
+  // "Network": "set using environment variable BFCLI_NETWORK",
+  // "ApiKey": "set using environment variable BFCLI_API_KEY",
+
+  // Advanced configuration
+  "Blockfrost": {
+    "Name": "Blockfrost.Net.Sdk-testnet",
+    "Network": "testnet",
+    "ConnectionLimit": 10
+    //"ApiKey": set through user secrets
+  }
+}
