+++ tools/Blockfrost.Api.Generate/Properties/launchSettings.json
@@ -2,7 +2,7 @@
   "profiles": {
     "Blockfrost.Api.Generate": {
       "commandName": "Project",
-      "commandLineArgs": "-s https://raw.githubusercontent.com/blockfrost/openapi/master/swagger.yaml -o C:\\dev\\blockfrost\\blockfrost-dotnet\\src\\Blockfrost.Api\\ -t C:\\dev\\blockfrost\\blockfrost-dotnet\\tools\\Blockfrost.Api.Generate\\templates_simple\\ -g v --preserve *LedgerService.cs,*MetadataService.cs,*NetworkService.cs"
+      "commandLineArgs": "-s https://raw.githubusercontent.com/blockfrost/openapi/master/swagger.yaml -o C:\\Users\\aklee\\source\\blockfrost\\blockfrost-dotnet\\src\\Blockfrost.Api\\ -t C:\\Users\\aklee\\source\\blockfrost\\blockfrost-dotnet\\tools\\Blockfrost.Api.Generate\\templates_simple\\ -g v --preserve *LedgerService.cs,*MetadataService.cs,*NetworkService.cs"
     }
   }
 }
