+++ tools/Blockfrost.Api.Generate/templates_simple/ABlockfrostService.hbr
@@ -14,6 +14,14 @@ namespace Blockfrost.Api
 {
     public partial class ABlockfrostService : IBlockfrostService
     {
+        private readonly Lazy<JsonSerializerOptions> _options;
+
+        private JsonSerializerOptions TextJsonSerializerSettings => _options.Value;
+
+        protected HttpClient HttpClient { get; set; }
+        
+        public string Name { get; internal set; }
+        
         public string Network { get; set; }
 
         public string BaseUrl
@@ -325,12 +333,6 @@ namespace Blockfrost.Api
 
             public string Text { get; }
         }
-
-        private readonly Lazy<JsonSerializerOptions> _options;
-
-        private JsonSerializerOptions TextJsonSerializerSettings => _options.Value;
-
-        protected HttpClient HttpClient { get; set; }
 
         private JsonSerializerOptions CreateSerializerOptions()
         {
