+++ tools/Blockfrost.Api.Generate/templates_simple/IBlockfrostService.hbr
@@ -10,6 +10,7 @@ namespace Blockfrost.Api
     public partial interface IBlockfrostService
     {
         string Network { get; }
+        string Name { get; }
         string BaseUrl { get; set; }
         bool ReadResponseAsString { get; set; }
     }
